requirements.txt
django>=3.0.5,<3.1.0
djangorestframework>=3.11.0,<3.12.0
flake8>=3.7.9,<3.8.0
psycopg2-binary>=2.8.4,<2.9.0